requirements.txt

streamlit>=1.36.0
python-dotenv>=1.0.1
openai>=1.44.0
pyyaml>=6.0.1
langgraph>=0.2.27
python-pptx>=0.6.23
